backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
pandas==2.1.3
numpy==1.26.2
ta==0.11.0
yfinance==0.2.32
matplotlib==3.8.2
scikit-learn==1.3.2
websockets==12.0
aiofiles==23.2.1

# Kiwoom API
PyQt5==5.15.10
pykiwoom==0.1.21
supabase==2.0.2

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0